refactor(topics): move topics folder to django_topics
