+++ django_topics/README.md
@@ -1,0 +1,8 @@
+# Django Topics app
+
+Django app that defines models and admin interfaces for editing certain aspects of topics that are unique to Sefaria's product and not needed for general usage of Sefaria's data.
+
+Currently contains methods to:
+- Edit which topics are in which pools
+- Define topic of the day schedule
+- Define seasonal topic schedule
