+++ templates/static/nash-bravmann-collection.html
@@ -110,7 +110,7 @@ text-decoration: underline;
             <span class="int-en">The Jack Nash and Ludwig Bravmann Collection is a free, <a
                     href="topics/adin-steinsaltz">online library of Rabbi Adin Even-Israel Steinsaltz's</a> major commentaries in Hebrew and English. Interlinked with Sefaria’s vast and ever-growing corpus of Jewish text, the Nash-Bravmann Collection further integrates Rabbi Steinsaltz’s Torah into the Jewish library, while also bringing these already-renowned commentaries to an even wider global audience.</span>
             <span class="int-he">האוסף על שם ג׳ק נאש ולודוויג ברוומן הינו
-                    <a href="/topics/adin-steinsaltz">
+                    <a href="/django_topics/adin-steinsaltz">
                     ספריה מקוונת חינמית הכוללת את הפרשנויות המרכזיות של הרב עדין אבן-ישראל שטיינזלץ</a>
                     בעברית ובאנגלית. בעצם שילובם של כתבי הרב שטיינזלץ באופן שוטף בספרייה ההולכת ומתרחבת של ספריא, אוסף נאש-ברוומן משמש כאמצעי להטמעה מעמיקה של תורתו בתוך שאר עולם המקורות היהודי, תוך כדי הנגשת הפרשנויות הנודעות האלו לקהילתנו העולמית.</span>
         </p>
